refactor: Move core package from `api.core` to `core`
Fixes #166
