+++ docling-core/src/main/java/ai/docling/core/package-info.java
@@ -2,6 +2,6 @@
  * Docling Core API.
  */
 @NullMarked
-package ai.docling.api.core;
+package ai.docling.core;
 
 import org.jspecify.annotations.NullMarked;
