+++ docling-core/src/test/java/ai/docling/core/DoclingDocumentTests.java
@@ -1,11 +1,11 @@
-package ai.docling.api.core;
+package ai.docling.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
-import ai.docling.api.core.DoclingDocument.DocItemLabel;
-import ai.docling.api.core.DoclingDocument.TitleItem;
+import ai.docling.core.DoclingDocument.DocItemLabel;
+import ai.docling.core.DoclingDocument.TitleItem;
 
 /**
  * Unit tests for {@link DoclingDocument}.
