+++ docling-serve/docling-serve-api/src/main/java/ai/docling/serve/api/chunk/response/ExportDocumentResponse.java
@@ -5,7 +5,7 @@ import org.jspecify.annotations.Nullable;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import ai.docling.api.core.DoclingDocument;
+import ai.docling.core.DoclingDocument;
 
 /**
  * Content of the original converted document, before getting chunked.
