+++ docling-serve/docling-serve-api/src/test/java/ai/docling/serve/api/convert/response/DocumentResponseTests.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import ai.docling.api.core.DoclingDocument;
-import ai.docling.api.core.DoclingDocument.DocItemLabel;
-import ai.docling.api.core.DoclingDocument.TitleItem;
+import ai.docling.core.DoclingDocument;
+import ai.docling.core.DoclingDocument.DocItemLabel;
+import ai.docling.core.DoclingDocument.TitleItem;
 
 /**
  * Unit tests for {@link DocumentResponse}.
